LayoutTests/imported/blink/http/tests/plugins/get-url-notify-on-removal-expected.txt

ALERT: Request completed
This tests that NPP_URLNotify gets called exactly once for canceled streams on plugin removal.
SUCCESS
